kb-qa-backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1 
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
werkzeug==3.0.3
zhipuai>=2.1.5
chromadb>=0.5.5
pypdf>=5.0.0
python-docx>=1.1.2
python-dotenv==1.0.1
sniffio>=1.3.0
